JsonDiff: factorize out null LCS node, part 2
No idea why this happens, but given the following conditions:

* LCS is not null;
* node1 is not null, and is not equivalent to the LCS node;
* node2 is not null, and is not equivalent to the LCS node;

then indices of both node1 and node2 must be equivalent.

This is an unforeseen consequence of testing nullity of LCS node first, and has
been detected by code coverage (no tests went through the path when all three
conditions held and indices were different).

Add a precondition, which appears to always be true. Now, it is left to
understand why that is.