Add serialization tests for JSON Patch operations
We know that we can successfully deserialize operations already using an
ObjectReader; now we test the ObjectWriter part.

These tests read the existing JSON test files, extract operations from them,
deserialize them and them serialize them again; we just have to test that the
JsonNodes produced by "reserialization" are the same as the original input.

We use JsonNumEquals for equality testing, since JSON equality in JSON Patch
matches that of JSON Schema.

With regards to issue #6, this ensures correct serialization of individual JSON
Patch operations; all that is left is JsonPatch itself.